requirements.txt
numpy
opencv-python
tqdm
Pillow
torch
torchvision
#rembg
wandb
segmentation-models
keras
tensorflow
glob
